+++ code/sys.ui/react.dev/src/Dev.render.tsx
@@ -1,5 +1,5 @@
 import { COLORS, DevWrangle, type t } from './common';
-import { DevKeyboard } from './f.Keyboard';
+import { DevKeyboard } from './fn.Keyboard';
 import { Harness } from './ui/Harness';
 import { ModuleList } from './ui/List.Module';
