--- code/sys.ui/react.dev/src/f.Bus/common.ts
@@ -1,2 +1,0 @@
-export * from '../common';
-export { Spec } from '../f.Spec';
