+++ code/sys.ui/react.dev/src/fn.Bus/Bus.Controller.mts
@@ -1,4 +1,4 @@
-import { Context } from '../f.Ctx';
+import { Context } from '../fn.Ctx';
 import { BusEvents } from './Bus.Events.mjs';
 import { BusMemoryState } from './Bus.MemoryState.mjs';
 import { DEFAULTS, Id, Is, R, Test, rx, type t } from './common';
