+++ code/sys.ui/react.dev/src/fn.Bus/Bus.MemoryState.mts
