+++ code/sys.ui/react.dev/src/fn.Bus/common.ts
@@ -1,5 +1,2 @@
 export * from '../common';
-export * from '../fn.Bus';
-export * from '../ui.use';
-
 export { Spec } from '../fn.Spec';
