+++ code/sys.ui/react.dev/src/fn.Bus/index.ts
