+++ code/sys.ui/react.dev/src/fn.Ctx/Ctx.PanelEdge.mts
