+++ code/sys.ui/react.dev/src/fn.Spec/Spec.ts
