+++ code/sys.ui/react.dev/src/test.ui/common.ts
@@ -1,5 +1,5 @@
 export { expect } from 'chai';
 export * from '../common';
 
-export { Spec } from '../f.Spec';
+export { Spec } from '../fn.Spec';
 export { TestLog } from './TestLog';
