+++ code/sys.ui/react.dev/src/test.ui/headless/headless.mts
@@ -1,6 +1,6 @@
 import 'symbol-observable';
 
-import { DevBus } from '../../f.Bus';
+import { DevBus } from '../../fn.Bus';
 import { Time, rx, slug, type t } from '../common';
 
 type Results = { ok: boolean; elapsed: t.Msecs; total: number; specs: ModuleResults[] };
